DimStandardJobRole repository, service, controller done

+++ src/MarketStat.Database/MarketStat.Database.Repositories/PostgresRepositories/Dimensions/DimStandardJobRoleHierarchyRepository.cs
@@ -1,8 +1,10 @@
 using MarketStat.Common.Converter.MarketStat.Common.Converter.Dimensions;
 using MarketStat.Common.Core.MarketStat.Common.Core.Dimensions;
+using MarketStat.Common.Exceptions;
 using MarketStat.Database.Context;
 using MarketStat.Database.Core.Repositories.Dimensions;
 using Microsoft.EntityFrameworkCore;
+using Npgsql;
 
 namespace MarketStat.Database.Repositories.PostgresRepositories.Dimensions;
 
@@ -18,13 +20,30 @@ public class DimStandardJobRoleHierarchyRepository : BaseRepository, IDimStandar
     {
         var dbLink = DimStandardJobRoleHierarchyConverter.ToDbModel(link);
         await _context.DimStandardJobRoleHierarchies.AddAsync(dbLink);
-        await _context.SaveChangesAsync();
+        try
+        {
+            await _context.SaveChangesAsync();
+        }
+        catch (DbUpdateException dbEx)
+            when (dbEx.InnerException is PostgresException pg
+                  && pg.SqlState == PostgresErrorCodes.UniqueViolation)
+        {
+            throw new ConflictException(
+                $"Link ({link.StandardJobRoleId}, {link.HierarchyLevelId}) already exists.");
+        }
+        catch (DbUpdateException dbEx)
+            when (dbEx.InnerException is PostgresException pg
+                  && pg.SqlState == PostgresErrorCodes.ForeignKeyViolation)
+        {
+            throw new NotFoundException(
+                $"Either StandardJobRole {link.StandardJobRoleId} or HierarchyLevel {link.HierarchyLevelId} does not exist.");
+        }
     }
 
     public async Task<DimStandardJobRoleHierarchy> GetStandardJobRoleHierarchyAsync(int jobRoleId, int levelId)
     {
-        var dbLink = await _context.DimStandardJobRoleHierarchies.FindAsync(jobRoleId, levelId)
-                     ?? throw new KeyNotFoundException($"Link ({jobRoleId}, {levelId}) not found.");
+        var dbLink = await _context.DimStandardJobRoleHierarchies.FindAsync(jobRoleId, levelId);
+        if (dbLink is null) throw new NotFoundException($"Link ({jobRoleId}, {levelId}) not found.");
         return DimStandardJobRoleHierarchyConverter.ToDomain(dbLink);
     }
 
@@ -52,8 +71,9 @@ public class DimStandardJobRoleHierarchyRepository : BaseRepository, IDimStandar
 
     public async Task DeleteStandardJobRoleHierarchyAsync(int jobRoleId, int levelId)
     {
-        var dbLink = await _context.DimStandardJobRoleHierarchies.FindAsync(jobRoleId, levelId)
-                     ?? throw new KeyNotFoundException($"Link ({jobRoleId}, {levelId}) not found.");
+        var dbLink = await _context.DimStandardJobRoleHierarchies.FindAsync(jobRoleId, levelId);
+        if (dbLink is null)
+            throw new NotFoundException($"Link ({jobRoleId}, {levelId}) not found.");
         _context.DimStandardJobRoleHierarchies.Remove(dbLink);
         await _context.SaveChangesAsync();
     }
